requirements.txt
catboost==0.24
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dill==0.3.2
Flask==1.1.2
graphviz==0.14.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
plotly==4.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1
